+++ Movit-Pi/src/movit-pi/Utils.h
@@ -44,7 +44,7 @@ struct imu_offset_t
 struct pressure_mat_offset_t
 {
     uint16_t analogOffset[PRESSURE_SENSOR_COUNT] = {0, 0, 0, 0, 0, 0, 0, 0, 0};
-    uint32_t totalSensorMean = 0;
+    float totalSensorMean = 0;
     float detectionThreshold = 0;
 };
